requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.4
python-multipart==0.0.9
Pillow==10.3.0
pytesseract==0.3.10
opencv-python==4.9.0.80
streamlit==1.35.0
pandas==2.2.2
matplotlib==3.9.0
seaborn==0.13.2
